+++ core/src/main/kotlin/net/corda/core/internal/notary/UniquenessProvider.kt
@@ -19,6 +19,7 @@ interface UniquenessProvider {
             txId: SecureHash,
             callerIdentity: Party,
             requestSignature: NotarisationRequestSignature,
-            timeWindow: TimeWindow? = null
+            timeWindow: TimeWindow? = null,
+            references: List<StateRef> = emptyList()
     )
 }