+++ core/src/main/kotlin/net/corda/core/transactions/LedgerTransaction.kt
@@ -42,7 +42,8 @@ data class LedgerTransaction @JvmOverloads constructor(
         override val notary: Party?,
         val timeWindow: TimeWindow?,
         val privacySalt: PrivacySalt,
-        private val networkParameters: NetworkParameters? = null
+        private val networkParameters: NetworkParameters? = null,
+        override val references: List<StateAndRef<ContractState>> = emptyList()
 ) : FullTransaction() {
     //DOCEND 1
     init {
@@ -66,10 +67,12 @@ data class LedgerTransaction @JvmOverloads constructor(
         }
     }
 
+    // Input reference state contracts are not required for verification.
     private val contracts: Map<ContractClassName, Try<Class<out Contract>>> = (inputs.map { it.state } + outputs)
             .map { it.contract to stateToContractClass(it) }.toMap()
 
     val inputStates: List<ContractState> get() = inputs.map { it.state.data }
+    val referenceStates: List<ContractState> get() = references.map { it.state.data }
 
     /**
      * Returns the typed input StateAndRef at the specified index
@@ -149,13 +152,13 @@ data class LedgerTransaction @JvmOverloads constructor(
 
     /**
      * Make sure the notary has stayed the same. As we can't tell how inputs and outputs connect, if there
-     * are any inputs, all outputs must have the same notary.
+     * are any inputs or reference inputs, all outputs must have the same notary.
      *
      * TODO: Is that the correct set of restrictions? May need to come back to this, see if we can be more
      *       flexible on output notaries.
      */
     private fun checkNoNotaryChange() {
-        if (notary != null && inputs.isNotEmpty()) {
+        if (notary != null && (inputs.isNotEmpty() || references.isNotEmpty())) {
             outputs.forEach {
                 if (it.notary != notary) {
                     throw TransactionVerificationException.NotaryChangeInWrongTransactionType(id, notary, it.notary)
@@ -255,6 +258,13 @@ data class LedgerTransaction @JvmOverloads constructor(
     fun getInput(index: Int): ContractState = inputs[index].state.data
 
     /**
+     * Helper to simplify getting an indexed reference input [ContractState].
+     * @param index the position of the item in the references.
+     * @return The [StateAndRef] at the requested index.
+     */
+    fun getReferenceInput(index: Int): ContractState = references[index].state.data
+
+    /**
      * Helper to simplify getting all inputs states of a particular class, interface, or base class.
      * @param clazz The class type used for filtering via an [Class.isInstance] check.
      * [clazz] must be an extension of [ContractState].
@@ -263,6 +273,16 @@ data class LedgerTransaction @JvmOverloads constructor(
     fun <T : ContractState> inputsOfType(clazz: Class<T>): List<T> = inputs.mapNotNull { clazz.castIfPossible(it.state.data) }
 
     inline fun <reified T : ContractState> inputsOfType(): List<T> = inputsOfType(T::class.java)
+
+    /**
+     * Helper to simplify getting all reference input states of a particular class, interface, or base class.
+     * @param clazz The class type used for filtering via an [Class.isInstance] check.
+     * [clazz] must be an extension of [ContractState].
+     * @return the possibly empty list of inputs matching the clazz restriction.
+     */
+    fun <T : ContractState> referenceInputsOfType(clazz: Class<T>): List<T> = references.mapNotNull { clazz.castIfPossible(it.state.data) }
+
+    inline fun <reified T : ContractState> referenceInputsOfType(): List<T> = referenceInputsOfType(T::class.java)
 
     /**
      * Helper to simplify getting all inputs states of a particular class, interface, or base class.
@@ -275,6 +295,18 @@ data class LedgerTransaction @JvmOverloads constructor(
     }
 
     inline fun <reified T : ContractState> inRefsOfType(): List<StateAndRef<T>> = inRefsOfType(T::class.java)
+
+    /**
+     * Helper to simplify getting all reference input states of a particular class, interface, or base class.
+     * @param clazz The class type used for filtering via an [Class.isInstance] check.
+     * [clazz] must be an extension of [ContractState].
+     * @return the possibly empty list of reference inputs [StateAndRef] matching the clazz restriction.
+     */
+    fun <T : ContractState> referenceInputRefsOfType(clazz: Class<T>): List<StateAndRef<T>> {
+        return references.mapNotNull { if (clazz.isInstance(it.state.data)) uncheckedCast<StateAndRef<ContractState>, StateAndRef<T>>(it) else null }
+    }
+
+    inline fun <reified T : ContractState> referenceInputRefsOfType(): List<StateAndRef<T>> = referenceInputRefsOfType(T::class.java)
 
     /**
      * Helper to simplify filtering inputs according to a [Predicate].
@@ -293,6 +325,22 @@ data class LedgerTransaction @JvmOverloads constructor(
     }
 
     /**
+     * Helper to simplify filtering reference inputs according to a [Predicate].
+     * @param clazz The class type used for filtering via an [Class.isInstance] check.
+     * [clazz] must be an extension of [ContractState].
+     * @param predicate A filtering function taking a state of type T and returning true if it should be included in the list.
+     * The class filtering is applied before the predicate.
+     * @return the possibly empty list of reference states matching the predicate and clazz restrictions.
+     */
+    fun <T : ContractState> filterReferenceInputs(clazz: Class<T>, predicate: Predicate<T>): List<T> {
+        return referenceInputsOfType(clazz).filter { predicate.test(it) }
+    }
+
+    inline fun <reified T : ContractState> filterReferenceInputs(crossinline predicate: (T) -> Boolean): List<T> {
+        return filterReferenceInputs(T::class.java, Predicate { predicate(it) })
+    }
+
+    /**
      * Helper to simplify filtering inputs according to a [Predicate].
      * @param predicate A filtering function taking a state of type T and returning true if it should be included in the list.
      * The class filtering is applied before the predicate.
@@ -306,6 +354,22 @@ data class LedgerTransaction @JvmOverloads constructor(
 
     inline fun <reified T : ContractState> filterInRefs(crossinline predicate: (T) -> Boolean): List<StateAndRef<T>> {
         return filterInRefs(T::class.java, Predicate { predicate(it) })
+    }
+
+    /**
+     * Helper to simplify filtering reference inputs according to a [Predicate].
+     * @param predicate A filtering function taking a state of type T and returning true if it should be included in the list.
+     * The class filtering is applied before the predicate.
+     * @param clazz The class type used for filtering via an [Class.isInstance] check.
+     * [clazz] must be an extension of [ContractState].
+     * @return the possibly empty list of references [StateAndRef] matching the predicate and clazz restrictions.
+     */
+    fun <T : ContractState> filterReferenceInputRefs(clazz: Class<T>, predicate: Predicate<T>): List<StateAndRef<T>> {
+        return referenceInputRefsOfType(clazz).filter { predicate.test(it.state.data) }
+    }
+
+    inline fun <reified T : ContractState> filterReferenceInputRefs(crossinline predicate: (T) -> Boolean): List<StateAndRef<T>> {
+        return filterReferenceInputRefs(T::class.java, Predicate { predicate(it) })
     }
 
     /**
@@ -326,6 +390,23 @@ data class LedgerTransaction @JvmOverloads constructor(
     }
 
     /**
+     * Helper to simplify finding a single reference inputs [ContractState] matching a [Predicate].
+     * @param clazz The class type used for filtering via an [Class.isInstance] check.
+     * [clazz] must be an extension of ContractState.
+     * @param predicate A filtering function taking a state of type T and returning true if this is the desired item.
+     * The class filtering is applied before the predicate.
+     * @return the single item matching the predicate.
+     * @throws IllegalArgumentException if no item, or multiple items are found matching the requirements.
+     */
+    fun <T : ContractState> findReference(clazz: Class<T>, predicate: Predicate<T>): T {
+        return referenceInputsOfType(clazz).single { predicate.test(it) }
+    }
+
+    inline fun <reified T : ContractState> findReference(crossinline predicate: (T) -> Boolean): T {
+        return findReference(T::class.java, Predicate { predicate(it) })
+    }
+
+    /**
      * Helper to simplify finding a single input matching a [Predicate].
      * @param clazz The class type used for filtering via an [Class.isInstance] check.
      * [clazz] must be an extension of ContractState.
@@ -340,6 +421,23 @@ data class LedgerTransaction @JvmOverloads constructor(
 
     inline fun <reified T : ContractState> findInRef(crossinline predicate: (T) -> Boolean): StateAndRef<T> {
         return findInRef(T::class.java, Predicate { predicate(it) })
+    }
+
+    /**
+     * Helper to simplify finding a single reference input matching a [Predicate].
+     * @param clazz The class type used for filtering via an [Class.isInstance] check.
+     * [clazz] must be an extension of ContractState.
+     * @param predicate A filtering function taking a state of type T and returning true if this is the desired item.
+     * The class filtering is applied before the predicate.
+     * @return the single item matching the predicate.
+     * @throws IllegalArgumentException if no item, or multiple items are found matching the requirements.
+     */
+    fun <T : ContractState> findReferenceInputRef(clazz: Class<T>, predicate: Predicate<T>): StateAndRef<T> {
+        return referenceInputRefsOfType(clazz).single { predicate.test(it.state.data) }
+    }
+
+    inline fun <reified T : ContractState> findReferenceInputRef(crossinline predicate: (T) -> Boolean): StateAndRef<T> {
+        return findReferenceInputRef(T::class.java, Predicate { predicate(it) })
     }
 
     /**
@@ -409,14 +507,26 @@ data class LedgerTransaction @JvmOverloads constructor(
      */
     fun getAttachment(id: SecureHash): Attachment = attachments.first { it.id == id }
 
-    fun copy(inputs: List<StateAndRef<ContractState>>,
-             outputs: List<TransactionState<ContractState>>,
-             commands: List<CommandWithParties<CommandData>>,
-             attachments: List<Attachment>,
-             id: SecureHash,
-             notary: Party?,
-             timeWindow: TimeWindow?,
-             privacySalt: PrivacySalt
-    ) = copy(inputs = inputs, outputs = outputs, commands = commands, attachments = attachments, id = id, notary = notary, timeWindow = timeWindow, privacySalt = privacySalt, networkParameters = null)
+    @JvmOverloads
+    fun copy(inputs: List<StateAndRef<ContractState>> = this.inputs,
+             outputs: List<TransactionState<ContractState>> = this.outputs,
+             commands: List<CommandWithParties<CommandData>> = this.commands,
+             attachments: List<Attachment> = this.attachments,
+             id: SecureHash = this.id,
+             notary: Party? = this.notary,
+             timeWindow: TimeWindow? = this.timeWindow,
+             privacySalt: PrivacySalt = this.privacySalt,
+             networkParameters: NetworkParameters? = this.networkParameters
+    ) = copy(inputs = inputs,
+            outputs = outputs,
+            commands = commands,
+            attachments = attachments,
+            id = id,
+            notary = notary,
+            timeWindow = timeWindow,
+            privacySalt = privacySalt,
+            networkParameters = networkParameters,
+            references = references
+    )
 }