+++ docs/source/release-notes.rst
@@ -4,6 +4,17 @@ Release notes
 Unreleased
 ----------
 
+Significant Changes in 4.0
+~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* **Reference states**:
+
+  Introduced the concept of "reference input states". A reference input state is a ``ContractState`` which can be referred
+  to in a transaction by the contracts of input and output states but whose contract is not executed as part of the
+  transaction verification process and is not consumed when the transaction is committed to the ledger but is checked
+  for "current-ness". In other words, the contract logic isn't run for the referencing transaction only. It's still a
+  normal state when it occurs in an input or output position.
+
 Issues Fixed
 ~~~~~~~~~~~~
 * Cordform Gradle task (`deployNodes`) doesn't work when `configFile` element was used.